requirements.txt
fastapi
uvicorn
sqlalchemy
passlib
bcrypt
python-jose
python-multipart
python-dotenv
psycopg2-binary==2.9.8
